Add files via upload
trees.m and species_data.m are the main two programs for this project; here is an explanation of what they both do:

trees.m 
This code ranks the best to worst growing conditions for carbon sequestration for each species.
There is an age field that can be changed which may change the ranking result. It has to be ran on each individual species file that I extracted from the WCC spreadsheet. 
It outputs the yield class and spacing which is optimal for carbon sequestration after a given amount of years and a given species.

species_data.m 
Once the data on which yield class and spacing was best for each species (after 25 years and 100 years) was found using trees.m, it was inputted into this code which produces two plots of the cumulative carbon sequestered per hectare after 25 and 100 years for the different species optimal growing conditions. This is so that the species can be compared with each other.